Added check for extension
Received files don't get deleted if there is an exception in
turn_file().  But if it succeeds, then the input file and the
turned one gets deleted.

Not only that but also, I save the input file to disc which
will overwrite whatever is there.  In that case, someone could
send me a file called server.jpg and that would delete overwrite
my image.

Now the file must end with .sgf.  And since there are no SGF files
that I want to keep, this is pretty good.

However, it it fails for a file, then I want to keep that file
so I might save that file as .sgf.failed and refuse to accept
files that failed.

I also need to impose a size limit.  It can be very generous, like
5Mb just to deal with attacks.