Added update-server.sh to automatically sync and restart the server with my machine
I do a git fetch and reset --hard to origin/master, then for uncommitted
changes, I do an rsync of the app/ directory.

I only rsync the app/ directory so that I can put files outside of it
and have them be different for each machine.  For example, the server
and the local machine must have different host.txt files which enable me
to have different behavior on each controlled by a file that is ignored
by git.